Split the TextUi to Formatter
